requirements.txt
PyYAML==5.1.2
pytest==5.1.0
codecov==2.0.15
